requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
audioread==3.0.1
bitsandbytes==0.45.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmudict==1.0.32
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gradio==5.26.0
gradio-client==1.9.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
kiwisolver==1.4.8
lazy-loader==0.4
lexicalrichness==0.5.1
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
mypy==1.15.0
mypy-extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.93.0
orjson==3.10.16
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pydub==0.25.1
pygments==2.19.1
pyparsing==3.2.3
pyphen==0.17.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.7
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.46.2
sympy==1.13.3
textblob==0.19.0
textstat==0.7.5
threadpoolctl==3.6.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typer==0.15.2
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
zipp==3.21.0
